Change from job scheduling form interval to cron.
When using `interval` for cron jobs the counter resets on every push. In
the event of multiple sequential pushes the counter will reset multiple
times and the execution of the job can be delayed significantly, delay
which will fire monitoring alerts.

Switching to cron scheduling with specific points in time to run the job
solves this issue.